+++ pkg/tasks/introspect.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/content-services/content-sources-backend/pkg/external_repos"
 	"github.com/content-services/content-sources-backend/pkg/models"
+	"github.com/content-services/content-sources-backend/pkg/tasks/payloads"
 	"github.com/content-services/content-sources-backend/pkg/tasks/queue"
 	"github.com/go-playground/validator/v10"
 	"github.com/rs/zerolog/log"
@@ -14,7 +15,7 @@ import (
 
 // TODO possibly remove context arg
 func IntrospectHandler(ctx context.Context, task *models.TaskInfo, _ *queue.Queue) error {
-	var p models.IntrospectPayload
+	var p payloads.IntrospectPayload
 	if err := json.Unmarshal(task.Payload, &p); err != nil {
 		return fmt.Errorf("payload incorrect type for IntrospectHandler")
 	}
