+++ pkg/tasks/payloads/repository_introspect.go
@@ -1,0 +1,8 @@
+package payloads
+
+const Introspect = "introspect"
+
+type IntrospectPayload struct {
+	Url   string
+	Force bool
+}
