+++ pkg/tasks/payloads/repository_snapshot.go
@@ -1,0 +1,10 @@
+package payloads
+
+const Snapshot = "snapshot"
+
+type SnapshotPayload struct {
+	SnapshotIdent        *string
+	SyncTaskHref         *string
+	PublicationTaskHref  *string
+	DistributionTaskHref *string
+}
